+++ product-service/src/main/java/com/eyamattoussi/productservice/dto/ProductRequestDto.java
@@ -4,6 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
@@ -11,5 +14,5 @@ import lombok.NoArgsConstructor;
 public class ProductRequestDto {
     private String name;
     private String description;
-    private Double price;
+    private BigDecimal price;
 }
